+++ src/App.css
@@ -1,4 +1,4 @@
-.addExpenseOrRevenueForm {
+/* .addExpenseOrRevenueForm {
   padding-top: 3em;
 }
 
@@ -36,4 +36,4 @@
   .roi-tables {
     padding-left: 1.5em;
   }
-}
+} */
