+++ src/components/Data.css
@@ -1,0 +1,40 @@
+.addExpenseOrRevenueForm {
+    padding-top: 3em;
+  }
+  
+  .input-field, .add-form-button {
+    padding-bottom: 1em;
+  }
+  
+  .roi-tables {
+    padding-top: 2em;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .expenses-table, .revenue-table, .totals-table {
+    margin-bottom: 3em;
+  }
+  .expenses-table th, .revenue-table th, .totals-table th {
+    width: 150px;
+  }
+  .expenses-table td, .revenue-table td, .totals-table td {
+    width: 100px
+  }
+  
+  .error {
+    color: red;
+  }
+  
+  @media only screen and (max-width:775px) {
+    .input-field, .add-form-button {
+      margin-left: .25em;
+      margin-right: .25em;
+    }
+    .roi-tables {
+      padding-left: 1.5em;
+    }
+  }
+
