+++ src/components/DataDisplay.js
@@ -1,0 +1,174 @@
+import React, { Component } from 'react';
+import {
+  Row,
+  Col,
+  Button,
+  Form
+} from 'react-bootstrap'
+import './Data.css';
+
+class DataDisplay extends Component {
+  constructor() {
+    super();
+    // "seed" data initially
+    this.state = {
+      revenue: [
+        {
+          name: 'Item 1',
+          oneTime: 100,
+          monthly: 50
+        },
+        {
+          name: 'Item 2',
+          oneTime: 50,
+          monthly: 25
+        },
+        {
+          name: 'Item 3',
+          oneTime: 25,
+          monthly: 85
+        }],
+      expenses: [{
+        name: 'Expense 1',
+        oneTime: 500,
+        monthly: 20.00
+      },
+      {
+        name: 'Expense 2',
+        oneTime: 200,
+        monthly: 40
+      }]
+    }
+
+    this.handleDelete = this.handleDelete.bind(this)
+    this.handleAdd = this.handleAdd.bind(this)
+  }
+
+  // Delete expense or revenue from list
+  handleDelete(type, index) {
+    // listType will be 'expenses' or 'revenue' depending on item to delete
+    let listType = this.state[type]
+    // recalculate and set totals in state
+    if (type === 'expenses') {
+      this.setState({
+        oneTimeExpense: this.state.oneTimeExpense - this.state.expenses[index]['oneTime'],
+        monthlyExpense: this.state.monthlyExpense - this.state.expenses[index]['monthly'],
+      })
+    } else {
+      // for revenue
+      this.setState({
+        oneTimeRevenue: this.state.oneTimeRevenue - this.state.revenue[index]['oneTime'],
+        monthlyRevenue: this.state.monthlyRevenue - this.state.revenue[index]['monthly'],
+      })
+    }
+    // remove list item from state array
+    this.setState({
+      [listType]: listType.splice(index, 1),
+    })
+  }
+
+  // add new expense or revenue
+  handleAdd(e) {
+    e.preventDefault()
+    // handle form errors, allows one-time and revenue amounts to be 0
+    if (!this.state.newType || !this.state.newName || (!this.state.newOneTime && this.state.newOneTime !== 0) || (!this.state.newMonthly && this.state.newMonthly !== 0)) {
+      this.setState({
+        error: true
+      })
+    }
+    // if there are no form errors, add accordingly
+    else {
+      // typeOfAmount will be either 'expenses' or 'revenue'
+      let typeOfAmount = this.state.newType
+      let monthly = typeOfAmount === 'expenses' ? 'monthlyExpense' : 'monthlyRevenue'
+      let oneTime = typeOfAmount === 'expenses' ? 'oneTimeExpense' : 'oneTimeRevenue'
+      // grab state array of revenues or expenses
+      let items = this.state[typeOfAmount]
+      items.push({
+        name: this.state.newName,
+        oneTime: this.state.newOneTime,
+        monthly: this.state.newMonthly
+      })
+      // set state with new totals and items array, clear errors displaying and form contents
+      this.setState({
+        error: false,
+        [typeOfAmount]: items,
+        [monthly]: this.state[monthly] + this.state.newMonthly,
+        [oneTime]: this.state[oneTime] + this.state.newOneTime,
+        //  Clear values in form
+        newName: '',
+        newMonthly: '',
+        newOneTime: '',
+        newType: ''
+      })
+    }
+  }
+
+  render() {
+    // create table rows from revenue state list
+    let revenueTableData = this.state.revenue.map((item, index) => {
+      return (
+        <tr key={"revenue" + index}>
+          <td>{item.name}</td>
+          <td>${item.oneTime.toFixed(2)}</td>
+          <td>${item.monthly.toFixed(2)}</td>
+          <td><Button onClick={() => this.handleDelete('revenue', index)}>Delete</Button></td>
+        </tr>
+      )
+    })
+    // create table rows from expenses state list
+    let expensesTableData = this.state.expenses.map((expense, index) => {
+      return (
+        <tr key={"expense" + index}>
+          <td>{expense.name}</td>
+          <td>${expense.oneTime.toFixed(2)}</td>
+          <td>${expense.monthly.toFixed(2)}</td>
+          <td><Button onClick={() => this.handleDelete('expenses', index)}>Delete</Button></td>
+        </tr>
+      )
+    })
+
+    return (
+      <div>
+        <div className="roi-tables">
+          {/* Revenue Table */}
+          <table className="revenue-table">
+            <thead>
+              <tr>
+                <th>Revenue</th>
+              </tr>
+              <tr>
+                <th></th>
+                <th>One-Time</th>
+                <th>Monthly</th>
+                <th></th>
+              </tr>
+            </thead>
+            <tbody>
+              {revenueTableData}
+            </tbody>
+          </table>
+          {/* Expenses Table */}
+          <table className="expenses-table">
+            <thead>
+              <tr>
+                <th>Expenses</th>
+              </tr>
+              <tr>
+                <th></th>
+                <th>One-Time</th>
+                <th>Monthly</th>
+                <th></th>
+              </tr>
+            </thead>
+            <tbody>
+              {expensesTableData}
+            </tbody>
+          </table>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default DataDisplay;
