+++ src/components/DataTotal.js
@@ -1,0 +1,89 @@
+import React, { Component } from 'react';
+import {
+  Row,
+  Col,
+  Button,
+  Form
+} from 'react-bootstrap'
+import './Data.css';
+
+class DataTotal extends Component {
+  constructor(props) {
+    super(props);
+    // "seed" data initially
+    this.state = {
+      oneTimeRevenue: 175,
+      oneTimeExpense: 700,
+      monthlyRevenue: 160,
+      monthlyExpense: 60
+    }
+  }
+
+  render() {
+    // MOVE THESE
+    // Calculations for totals
+    let totalRevenue = this.state.oneTimeRevenue + (this.state.monthlyRevenue * 24)
+    let totalExpense = this.state.oneTimeExpense + (this.state.monthlyExpense * 24)
+    let monthlyContributionProfit = this.state.monthlyRevenue - this.state.monthlyExpense
+    let totalContributionProfit = totalRevenue - totalExpense
+    // handle case where totalRevenue is 0 (to avoid -Infinity and NaN)
+    let contributionMargin = totalRevenue !== 0 ? (totalContributionProfit / totalRevenue * 100).toFixed(0) : 0
+    // handle case where totalExpense and totalRevenue are 0 (to avoid NaN)
+    let capitalROI = (totalExpense === 0 && totalRevenue === 0) ? 0 : ((this.state.oneTimeExpense - this.state.oneTimeRevenue) / monthlyContributionProfit).toFixed(1)
+    
+    return (
+      <div>
+        <div className="roi-tables">
+          {/* Totals Table */}
+          <table className="totals-table">
+            <thead>
+              <tr>
+                <th>24 Month Term</th>
+              </tr>
+              <tr>
+                <th>Total</th>
+                <th>One-Time</th>
+                <th>Monthly</th>
+                <th>Total</th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr>
+                <td>Revenue</td>
+                <td>${(this.state.oneTimeRevenue).toFixed(2)}</td>
+                <td>${(this.state.monthlyRevenue).toFixed(2)}</td>
+                <td>${totalRevenue.toFixed(2)}</td>
+              </tr>
+              <tr>
+                <td>Expenses</td>
+                <td>${(this.state.oneTimeExpense).toFixed(2)}</td>
+                <td>${(this.state.monthlyExpense).toFixed(2)}</td>
+                <td>${totalExpense.toFixed(2)}</td>
+              </tr>
+              <tr>
+                <td>Contribution Profit</td>
+                <td></td>
+                <td>${monthlyContributionProfit.toFixed(2)}</td>
+                <td>${totalContributionProfit.toFixed(2)}</td>
+              </tr>
+              <tr>
+                <td>Contribution Margin</td>
+                <td></td>
+                <td></td>
+                <td>{contributionMargin}%</td>
+              </tr>
+              <tr>
+                <td>Capital ROI (monthly)</td>
+                <td></td>
+                <td></td>
+                <td>{capitalROI}</td>
+              </tr>
+            </tbody>
+          </table>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default DataTotal;
