+++ src/data/seedData.js
@@ -1,0 +1,38 @@
+export default [{
+    revenue: [
+    {
+      name: 'Item 1',
+      oneTime: 100,
+      monthly: 50
+    },
+    {
+      name: 'Item 2',
+      oneTime: 50,
+      monthly: 25
+    },
+    {
+      name: 'Item 3',
+      oneTime: 25,
+      monthly: 85
+    }],
+    expenses:[{
+      name: 'Expense 1',
+      oneTime: 500,
+      monthly: 20.00
+    },
+    {
+      name: 'Expense 2',
+      oneTime: 200,
+      monthly: 40
+    }],
+    oneTimeRevenue: 175,
+    oneTimeExpense: 700,
+    monthlyRevenue: 160,
+    monthlyExpense: 60,
+    newType: '',
+    newName: '',
+    newOneTime: '',
+    newMonthly: '',
+    error: false
+  }];
+
